Can now continue on existing map directory

+++ main.cpp
@@ -3,6 +3,7 @@
 #include <fstream>
 #include <sstream>
 #include <vector>
+#include <set>
 #include <filesystem>
 
 #include "Path.h"
@@ -13,9 +14,9 @@
 using namespace std;
 namespace fs = std::filesystem;
 
-int main(int argc, char* argv[]) {
+int main(int argc, char const * argv[]) {
     //Make sure path is specified
-    if (argc != 2) {
+    if (argc < 2) {
         cout << "Incorrect number of command line arguments." << endl;
         return 0;
     }
@@ -24,13 +25,39 @@ int main(int argc, char* argv[]) {
     ifstream inFile;
     ofstream outFile;
     fs::path currentPath = fs::current_path();
-    Path startFile(argv[1]);
+    Path startFile;
 
-    //Remove previous path map
-    fs::remove_all("Map");
+    if (argc == 2) {
+        cout << "Starting new map!" << endl;
 
-    //Create new Map directory
-    fs::create_directories("Map");
+        startFile = Path(argv[1]);
+
+        //Remove previous path map
+        fs::remove_all("Map");
+
+        //Create new Map directory
+        fs::create_directories("Map");
+    }
+    else if (argc == 3 && (string)argv[1] == "-c") {
+        cout << "Continuing on existing map!" << endl;
+
+        startFile = Path(argv[2]);
+
+        //Load FilesVisited.txt into filesVisited
+        inFile.open("FilesVisited.txt");
+        if (inFile.is_open()) {
+            string inputLine;
+            while (getline(inFile, inputLine)) {
+                filesVisited.insert(inputLine);
+            }
+        }
+        inFile.close();
+    }
+    else {
+        cout << "Incorrect command line argument." << endl;
+        return 0;
+    }
+    
 
     //Verify startFile
     inFile.open(startFile.GetAbsolutePath());
@@ -48,6 +75,13 @@ int main(int argc, char* argv[]) {
 
     cout << "\n\nPath-Finder Mapping Complete!" << endl;
     cout << filesVisited.size() << " files read!" << endl;
+    outFile.open("FilesVisited.txt");
+    if (outFile.is_open()) {
+        for (fs::path pa : filesVisited) {
+            outFile << pa << endl;
+        }
+    }
+    outFile.close();
     cout << pathsMade.size() << " paths/files made!" << endl;
     return 0;
 }
